# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ACROSS Server",
        "summary": "Astrophysics Cross-Observatory Science Support (ACROSS)",
        "description": "Server providing tools and utilities for various NASA missions to aid in coordination of large observation efforts.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tle/": {
            "post": {
                "tags": [
                    "TLE"
                ],
                "summary": "Create a TLE",
                "description": "Create a new TLE for ACROSS.",
                "operationId": "create_tle",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TLECreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TLE"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The TLE does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/telescope/{telescope_id}": {
            "get": {
                "tags": [
                    "Telescope"
                ],
                "summary": "Read an telescope",
                "description": "Read a telescope by a telescope ID.",
                "operationId": "get_telescope",
                "parameters": [
                    {
                        "name": "telescope_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Telescope Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a Telescope",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Telescope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Telescope not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/telescope/": {
            "get": {
                "tags": [
                    "Telescope"
                ],
                "summary": "Read telescopes(s)",
                "description": "Read most recent telescopes based on query params",
                "operationId": "get_telescopes",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Name"
                        }
                    },
                    {
                        "name": "instrument_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Instrument Id"
                        }
                    },
                    {
                        "name": "instrument_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Instrument Name"
                        }
                    },
                    {
                        "name": "created_on",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Created On"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a telescope",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Telescope"
                                    },
                                    "title": "Response 200 Get Telescopes"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The telescope does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IDNameSchema": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "IDNameSchema"
            },
...
            "TLE": {
                "properties": {
                    "norad_id": {
                        "type": "integer",
                        "title": "Norad Id"
                    },
                    "satellite_name": {
                        "type": "string",
                        "title": "Satellite Name"
                    },
                    "tle1": {
                        "type": "string",
                        "maxLength": 69,
                        "minLength": 69,
                        "title": "Tle1"
                    },
                    "tle2": {
                        "type": "string",
                        "maxLength": 69,
                        "minLength": 69,
                        "title": "Tle2"
                    },
                    "epoch": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Epoch",
                        "description": "Calculate the Epoch of the TLE file. See\nhttps://celestrak.org/columns/v04n03/#FAQ04 for more information on how\nthe year / epoch encoding works.\n\nReturns\n-------\n    The calculated epoch of the TLE.",
                        "readOnly": true
                    }
                },
                "type": "object",
                "required": [
                    "norad_id",
                    "satellite_name",
                    "tle1",
                    "tle2",
                    "epoch"
                ],
                "title": "TLE"
            },
            "TLECreate": {
                "properties": {
                    "norad_id": {
                        "type": "integer",
                        "title": "Norad Id"
                    },
                    "satellite_name": {
                        "type": "string",
                        "title": "Satellite Name"
                    },
                    "tle1": {
                        "type": "string",
                        "maxLength": 69,
                        "minLength": 69,
                        "title": "Tle1"
                    },
                    "tle2": {
                        "type": "string",
                        "maxLength": 69,
                        "minLength": 69,
                        "title": "Tle2"
                    }
                },
                "type": "object",
                "required": [
                    "norad_id",
                    "satellite_name",
                    "tle1",
                    "tle2"
                ],
                "title": "TLECreate"
            },
...
            "Telescope": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "created_on": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created On"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "short_name": {
                        "type": "string",
                        "title": "Short Name"
                    },
                    "observatory": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/IDNameSchema"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "instruments": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/IDNameSchema"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Instruments"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "created_on",
                    "name",
                    "short_name"
                ],
                "title": "Telescope",
                "description": "A Pydantic model class representing a created telescope\n\nNotes\n-----\nInherits from TelescopeBase\n\nMethods\n-------\nfrom_orm(telescope: TelescopeModel) -> Telescope\n    Static method that instantiates this class from a telescope database record"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Enter your access token.",
                "scheme": "bearer"
            }
        }
    }
}